Saving is now supported. (#25)
CTRL+S saves to the current file.
If no file was specified, a prompt for the file name will appear.

CTRL+ALT+S Will always prompt for the file name. If a file name was
Implemneted SAVE and SAVE AS.
already chosen, it shows up on the prompt.

CTRL+Q now asks if you want to save in case the buffer has changed.